# docs/reconstruction/assets/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateCaseRequest": {
        "additionalProperties": false,
        "description": "Backend API payload; Backend creates the case and starts/resumes the graph.",
        "properties": {
          "attached_artifact_refs": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "title": "Attached Artifact Refs",
            "type": "array"
          },
          "initial_message": {
            "minLength": 1,
            "title": "Initial Message",
            "type": "string"
          },
          "order_ref": {
            "minLength": 1,
            "title": "Order Ref",
            "type": "string"
          },
          "user_ref": {
            "minLength": 1,
            "title": "User Ref",
            "type": "string"
          }
        },
        "required": [
          "order_ref",
          "user_ref",
          "initial_message"
        ],
        "title": "CreateCaseRequest",
        "type": "object"
      },
      "CreateCaseResponse": {
        "additionalProperties": false,
        "properties": {
          "case_ref": {
            "minLength": 1,
            "title": "Case Ref",
            "type": "string"
          }
        },
        "required": [
          "case_ref"
        ],
        "title": "CreateCaseResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Adaptive Return Resolution API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/cases": {
      "post": {
        "operationId": "create_case_cases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Case"
      }
    },
    "/cases/{case_ref}": {
      "get": {
        "operationId": "get_case_cases__case_ref__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_ref",
            "required": true,
            "schema": {
              "title": "Case Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Case"
      }
    },
...
    "/cases/{case_ref}/events": {
      "get": {
        "operationId": "stream_case_events_cases__case_ref__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_ref",
            "required": true,
            "schema": {
              "title": "Case Ref",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Case Events"
      }
    },
    "/cases/{case_ref}/messages": {
      "post": {
        "operationId": "send_message_cases__case_ref__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_ref",
            "required": true,
            "schema": {
              "title": "Case Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Message"
      }
    },
    "/cases/{case_ref}/review": {
      "post": {
        "operationId": "complete_human_review_cases__case_ref__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_ref",
            "required": true,
            "schema": {
              "title": "Case Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "APPROVE": "#/components/schemas/ApproveReviewDecision",
                    "EDIT": "#/components/schemas/EditReviewDecision",
                    "REJECT": "#/components/schemas/RejectReviewDecision"
                  },
                  "propertyName": "decision"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/ApproveReviewDecision"
                  },
                  {
                    "$ref": "#/components/schemas/EditReviewDecision"
                  },
                  {
                    "$ref": "#/components/schemas/RejectReviewDecision"
                  }
                ],
                "title": "Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Human Review"
      }
    },
...
    "/internal/v1/case-context": {
      "post": {
        "operationId": "load_case_context_internal_v1_case_context_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoadCaseContextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadCaseContextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Load Case Context"
      }
    },
    "/internal/v1/evidence/resolve": {
      "post": {
        "operationId": "resolve_evidence_internal_v1_evidence_resolve_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveEvidenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveEvidenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Evidence"
      }
    },
    "/internal/v1/human-reviews": {
      "post": {
        "operationId": "submit_human_review_internal_v1_human_reviews_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitHumanReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitHumanReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Human Review"
      }
    },
    "/internal/v1/human-reviews/result": {
      "post": {
        "operationId": "fetch_human_review_result_internal_v1_human_reviews_result_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FetchHumanReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchHumanReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fetch Human Review Result"
      }
    },
    "/internal/v1/memory/candidates": {
      "post": {
        "operationId": "submit_memory_candidate_internal_v1_memory_candidates_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitMemoryCandidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitMemoryCandidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Memory Candidate"
      }
    },
...
  }
}